@@ -37,11 +37,11 @@ public extension Promise {
 	/// Catch a cancelled promise.
 	///
 	/// - Parameters:
-	///   - context: context in which the body will be eecuted. If not specified `.main` is used.
+	///   - context: context in which the body will be executed. If not specified `.main` is used.
 	///   - body: body to execute
 	/// - Returns: a new void promise
 	@discardableResult
-	func cancelled(in context: Context? = nil, _ body: @escaping (() -> (()))) -> Promise<Void> {
+	func cancelled(in context: Context? = nil, _ body: @escaping (() -> (Void))) -> Promise<Void> {
 		let ctx = context ?? .main
 		let nextPromise = Promise<Void>(in: ctx, token: self.invalidationToken) { resolve, reject, operation in
 			let onResolve = Observer.onResolve(ctx, { _ in
